Refactor assessment wizard UI tests
